Splitting up even more and removing old specs
