--- Specifications/for_BindingsPerTenant/given/type_binding.cs
@@ -1,25 +1,0 @@
-using Autofac;
-using Machine.Specifications;
-using Moq;
-
-namespace Dolittle.DependencyInversion.Autofac.Specifications.for_BindingsPerTenant.given
-{
-    public class type_binding : bindings_per_tenants
-    {
-        protected interface first_dependency {}
-        protected interface second_dependency {}
-
-        protected class some_type
-        {
-        }
-
-        protected static Binding binding;
-        protected static IComponentContext component_context;
-
-        Establish context = () =>
-        {
-            component_context = Mock.Of<IComponentContext>();
-            binding = new Binding(typeof(some_type), new Strategies.Type(typeof(some_type)), new Scopes.SingletonPerTenant());
-        };
-    }
-}
