+++ Specifications/for_InstancesPerTenant/when_resolving_type_binding_twice_for_same_tenant.cs
@@ -1,0 +1,7 @@
+namespace Specifications.for_InstancesPerTenant
+{
+    public class when_resolving_type_binding_twice_for_same_tenant
+    {
+        
+    }
+}
